# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autoken Facturas v2",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "intake"
        ],
        "summary": "Upload File",
        "description": "Sube un fichero de factura a una empresa. Ver spec S2.1 para los c\u00f3digos (201/4xx/503).\n\nOrden: pertenencia (403/404) -> tama\u00f1o (413) -> el servicio hace el resto (415/422/409/503/201).\n\n`sharpness_score` (S6.14 C8): nitidez de la captura calculada en cliente (varianza del\nLaplaciano), string opcional. Es TELEMETR\u00cdA, no dato de dominio: no se persiste ni se valida\n(un valor raro no rompe la subida); solo se loguea como m\u00e9trica de calidad de captura.",
        "operationId": "upload_file_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/batch": {
      "post": {
        "tags": [
          "intake"
        ],
        "summary": "Upload Batch",
        "description": "Acepta de dos a cinco im\u00e1genes como un \u00fanico documento multip\u00e1gina (S6.12).\n\nLa direcci\u00f3n se valida aqu\u00ed porque forma parte del contrato de captura. La confirmaci\u00f3n\nexistente sigue siendo quien la persiste, igual que en una subida simple.\n\n`sharpness_score` (S6.14 C8): misma telemetr\u00eda opcional que en la subida simple (hoy el cliente\nno calcula una nitidez de conjunto para varias p\u00e1ginas; el campo se admite igual).",
        "operationId": "upload_batch_api_v1_uploads_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_batch_api_v1_uploads_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/status": {
      "get": {
        "tags": [
          "intake"
        ],
        "summary": "Upload Status",
        "description": "Consulta el estado de OCR con la misma autorizaci\u00f3n privada que la descarga.",
        "operationId": "upload_status_api_v1_uploads__file_id__status_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/retry-ocr": {
      "post": {
        "tags": [
          "intake"
        ],
        "summary": "Retry Ocr",
        "description": "Reencola de forma autorizada una lectura que termin\u00f3 en fallo, sin volver a subir bytes.",
        "operationId": "retry_ocr_api_v1_uploads__file_id__retry_ocr_post",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrRetryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/download-url": {
      "get": {
        "tags": [
          "intake"
        ],
        "summary": "Download Url",
        "description": "URL firmada (5 min) para descargar un fichero de intake. Ver spec S2.7 para los c\u00f3digos.\n\nAutorizaci\u00f3n id\u00e9ntica a `review`/`confirm` (S2.5): 403 empresa hermana del propio tenant, 404\notro tenant/inexistente. La descarga no depende del estado del fichero (spec S2.7 \u00a75).",
        "operationId": "download_url_api_v1_uploads__file_id__download_url_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUrlOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/image": {
      "get": {
        "tags": [
          "intake"
        ],
        "summary": "Download Image",
        "description": "Bytes reales del fichero de intake (2026-08-01), v\u00eda la API \u2014 no una redirecci\u00f3n a MinIO.\n\nReemplaza a `download-url` como camino real del bot\u00f3n \"Ver\" del panel: MinIO nunca se expone\np\u00fablicamente en este proyecto, as\u00ed que una URL firmada de MinIO es inalcanzable desde el\nnavegador del usuario en el despliegue real (ver docstring de `service.get_download_bytes`).\nMisma autorizaci\u00f3n que `download-url` (403/404).",
        "operationId": "download_image_api_v1_uploads__file_id__image_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bytes binarios del documento original autorizado.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{file_id}/pages/{page_number}/image": {
      "get": {
        "tags": [
          "intake"
        ],
        "summary": "Download Page Image",
        "description": "Bytes de una hoja secundaria, con la misma privacidad por usuario que la ra\u00edz (S6.12).",
        "operationId": "download_page_image_api_v1_uploads__file_id__pages__page_number__image_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          },
          {
            "name": "page_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bytes binarios de la imagen original autorizada.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invoices/{invoice_id}": {
      "patch": {
        "tags": [
          "invoicing"
        ],
        "summary": "Edit Invoice",
        "description": "Edita los campos presentes en el body de una factura confirmada. Solo `tenant_admin` (S3.3).\n\nPatch parcial real: solo las claves que el cliente envi\u00f3 (`model_fields_set`) llegan al\nservicio; un campo ausente en el body conserva su valor actual (spec \u00a72). `tax_lines: null`\nexpl\u00edcito se trata como AUSENTE (no toca los tramos): solo una lista real (incluida `[]`, que\nlos borra todos) es una instrucci\u00f3n de cambio; un `null` no es una lista de tramos v\u00e1lida.",
        "operationId": "edit_invoice_api_v1_invoices__invoice_id__patch",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceEditIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/platform/tenants/{tenant_id}/invoices": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List Tenant Invoices",
        "description": "Facturas confirmadas del tenant elegido desde el laboratorio (S6.2, spec C2). Id\ninexistente -> 404 expl\u00edcito (spec C4).",
        "operationId": "list_tenant_invoices_api_v1_platform_tenants__tenant_id__invoices_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabInvoiceRowOut"
                  },
                  "title": "Response List Tenant Invoices Api V1 Platform Tenants  Tenant Id  Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_v1_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "recibida",
                  "emitida"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "sharpness_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpness Score"
          }
        },
        "type": "object",
        "required": [
          "file",
          "company_id"
        ],
        "title": "Body_upload_file_api_v1_uploads_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OcrRetryOut": {
        "properties": {
          "status": {
            "type": "string",
            "const": "pending_ocr",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OcrRetryOut"
      },
...
    }
  }
}